FIX: Changed 2nd parameter of check_shape() from unsigned int * to size_t *.
git-svn-id: http://svn.boost.org/svn/boost/trunk@17831 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/multi_array/test/constructors.cpp
@@ -10,12 +10,12 @@
 #include <algorithm>
 #include <list>
 
-void check_shape(const double&, unsigned int*, int*, unsigned int)
+void check_shape(const double&, size_t*, int*, unsigned int)
 {}
 
 template <class Array>
 void check_shape(const Array& A, 
-                 unsigned int* sizes, 
+                 size_t* sizes, 
                  int* strides,
                  unsigned int num_elements)
 {
